# controlplane/gen/.openapi-3.0.yaml
openapi: 3.0.3
info:
  title: LangSmith Deployment
  version: 0.1.0
  description: "LangGraph Cloud-compatible API surface implemented by duragraph.\n\
    \nNote: paths are documented as bare (`/assistants`, `/threads`, `/runs`,\netc.)\
    \ to mirror the upstream LangGraph spec. The duragraph engine\nserves these paths\
    \ under the `/api/v1` base \u2014 declared via the\n`servers:` block below. Clients\
    \ should construct request URLs as\n`<server.url><path>`, e.g. `http://localhost:8081/api/v1/runs`.\n"
servers:
- url: http://localhost:8081/api/v1
  description: Local development server
- url: https://api.duragraph.ai/api/v1
  description: "Production server (placeholder \u2014 replace with deployment host)"
...
paths:
...
  /threads/{thread_id}/runs/crons:
    post:
      tags:
      - Crons (Plus tier)
      summary: Create Thread Cron
      description: Create a cron to schedule runs on a thread.
      operationId: create_thread_cron_threads__thread_id__runs_crons_post
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CronCreate'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cron'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /runs/crons:
    post:
      tags:
      - Crons (Plus tier)
      summary: Create Cron
      description: Create a cron to schedule runs on new threads.
      operationId: create_cron_runs_crons_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CronCreate'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cron'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /runs/crons/search:
    post:
      tags:
      - Crons (Plus tier)
      summary: Search Crons
      description: Search all active crons
      operationId: search_crons_runs_crons_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CronSearch'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Cron'
                type: array
                title: Response Search Crons Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /runs/crons/count:
    post:
      tags:
      - Crons (Plus tier)
      summary: Count Crons
      description: Get the count of crons matching the specified criteria.
      operationId: count_crons_runs_crons_count_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CronCountRequest'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: integer
                title: Count
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /runs/crons/{cron_id}:
    delete:
      tags:
      - Crons (Plus tier)
      summary: Delete Cron
      description: Delete a cron by ID.
      operationId: delete_cron_runs_crons__cron_id__delete
      parameters:
      - required: true
        schema:
          type: string
          format: uuid
          title: Cron Id
        name: cron_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Cron:
      properties:
        cron_id:
          type: string
          format: uuid
          title: Cron Id
          description: The ID of the cron.
        assistant_id:
          type: string
          format: uuid
          title: Assistant Id
          description: The ID of the assistant.
          nullable: true
        thread_id:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        end_time:
          type: string
          format: date-time
          title: End Time
          description: The end date to stop running the cron.
        schedule:
          type: string
          title: Schedule
          description: The schedule to run, cron format.
        created_at:
          type: string
          format: date-time
          title: Created At
          description: The time the cron was created.
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: The last time the cron was updated.
        user_id:
          type: string
          title: User Id
          description: The ID of the user.
          nullable: true
        payload:
          type: object
          title: Payload
          description: The run payload to use for creating new run.
        next_run_date:
          type: string
          format: date-time
          title: Next Run Date
          description: The next run date of the cron.
          nullable: true
        metadata:
          type: object
          title: Metadata
          description: The cron metadata.
      type: object
      required:
      - cron_id
      - thread_id
      - end_time
      - schedule
      - created_at
      - updated_at
      - payload
      title: Cron
      description: Represents a scheduled task.
    CronCreate:
      properties:
        schedule:
          type: string
          title: Schedule
          description: The cron schedule to execute this job on.
        end_time:
          type: string
          format: date-time
          title: End Time
          description: The end date to stop running the cron.
        assistant_id:
          description: The assistant ID or graph name to run. If using graph name,
            will default to the assistant automatically created from that graph by
            the server.
        input:
          title: Input
          description: The input to the graph.
        metadata:
          type: object
          title: Metadata
          description: Metadata to assign to the cron job runs.
        config:
          properties:
            tags:
              items:
                type: string
              type: array
              title: Tags
            recursion_limit:
              type: integer
              title: Recursion Limit
            configurable:
              type: object
              title: Configurable
          type: object
          title: Config
          description: The configuration for the assistant.
        context:
          type: object
          title: Context
          description: Static context added to the assistant.
        webhook:
          type: string
          maxLength: 65536
          minLength: 1
          format: uri-reference
          title: Webhook
          description: Webhook to call after LangGraph API call is done.
        interrupt_before:
          title: Interrupt Before
          description: Nodes to interrupt immediately before they get executed.
        interrupt_after:
          title: Interrupt After
          description: Nodes to interrupt immediately after they get executed.
        multitask_strategy:
          type: string
          enum:
          - reject
          - rollback
          - interrupt
          - enqueue
          title: Multitask Strategy
          description: Multitask strategy to use. Must be one of 'reject', 'interrupt',
            'rollback', or 'enqueue'.
          default: enqueue
      type: object
      required:
      - assistant_id
      - schedule
      title: CronCreate
      description: Payload for creating a cron job.
    CronSearch:
      properties:
        assistant_id:
          type: string
          format: uuid
          title: Assistant Id
          description: The assistant ID or graph name to filter by using exact match.
        thread_id:
          type: string
          format: uuid
          title: Thread Id
          description: The thread ID to search for.
        limit:
          type: integer
          title: Limit
          description: The maximum number of results to return.
          default: 10
          minimum: 1
          maximum: 1000
        offset:
          type: integer
          title: Offset
          description: The number of results to skip.
          default: 0
          minimum: 0
        sort_by:
          type: string
          title: Sort By
          description: The field to sort by.
          default: created_at
          enum:
          - cron_id
          - assistant_id
          - thread_id
          - next_run_date
          - end_time
          - created_at
          - updated_at
        sort_order:
          type: string
          title: Sort Order
          description: The order to sort by.
          default: desc
          enum:
          - asc
          - desc
        select:
          type: array
          items:
            type: string
            enum:
            - cron_id
            - assistant_id
            - thread_id
            - end_time
            - schedule
            - created_at
            - updated_at
            - user_id
            - payload
            - next_run_date
            - metadata
            - now
          title: Select
          description: Specify which fields to return. If not provided, all fields
            are returned.
      type: object
      required: []
      title: CronSearch
      description: Payload for listing crons
    CronCountRequest:
      properties:
        assistant_id:
          type: string
          format: uuid
          title: Assistant Id
          description: The assistant ID or graph name to search for.
        thread_id:
          type: string
          format: uuid
          title: Thread Id
          description: The thread ID to search for.
      type: object
      title: CronCountRequest
      description: Payload for counting crons
...
    ErrorResponse:
      type: string
      title: ErrorResponse
      description: Error message returned from the server
...